+++ utils/misc.py
@@ -2,15 +2,20 @@ import subprocess
 import os
 import sys
 import shutil
-import threading
-import tkinter as tk
+import re
+from datetime import datetime
+from math import floor
 from typing import Any
 from functools import reduce
 import operator
+import threading
 
-"""
-  Class decorators
-"""
+import tkinter as tk
+
+from utils.debug import Debug, catch_exceptions
+
+
+""" Class decorators """
 def singleton(cls):
     """ A thread-safe implementation of Singleton. Note this will break unittest.mock.patch """
     instances = {}
@@ -26,9 +31,7 @@ def singleton(cls):
 
     return get_instance
 
-"""
-  Miscellaneous utility functions
-"""
+""" Miscellaneous utility functions """
 def get_by_path(dic:dict[str, Any], keys:list[str], default:Any = None) -> Any:
     """ Return an element from a nested object by item sequence. """
     try:
@@ -36,8 +39,9 @@ def get_by_path(dic:dict[str, Any], keys:list[str], default:Any = None) -> Any:
     except (KeyError, IndexError, TypeError):
         return default
 
+@catch_exceptions
 def copy_to_clipboard(parent:tk.Widget|None, text:str = '') -> None:
-    """ Copy text to the clipboard o windows or Linx, X11 or Wayland """
+    """ Copy text to the clipboard """
     if parent == None: return
 
     # Non-linux is easy.
@@ -61,26 +65,168 @@ def copy_to_clipboard(parent:tk.Widget|None, text:str = '') -> None:
                 break
 
     if clipboard_cli != None:
+        Debug.logger.debug(f"Using linux clipboard: {clipboard_cli}")
         try:
             subprocess.run(clipboard_cli.split(), input=text.encode('utf-8'), check=True)
         except subprocess.CalledProcessError as e:
-            pass
+            Debug.logger.error(f"Failed to run {clipboard_cli}: {e}")
         return
 
     # Still nothing? Then run all the ones we can find regardless of session type.
     for cmd in cmds:
         if shutil.which(cmd.split()[0]):
-            clipboard_cli = cmd
+            cli:str = cmd
             try:
-                if clipboard_cli != None:
-                    subprocess.run(clipboard_cli.split(), input=text.encode('utf-8'), check=True)
+                subprocess.run(cli.split(), input=text.encode('utf-8'), check=True)
             except subprocess.CalledProcessError as e:
-                pass
+                Debug.logger.error(f"Failed to run {cli}: {e}")
 
     if clipboard_cli != None:
         return
 
     # Final fallback to the tkinter version
+    Debug.logger.debug(f"Using linux tkinter clipboard fallback")
     parent.clipboard_clear()
     parent.clipboard_append(text)
     parent.update()
+
+
+def hfplus(val:int|float|str|bool|tuple, type:str|None = None) -> str:
+    """
+        A general customized formatting function.
+        Args:
+            val (int|float|str|bool|tuple): A tuple or a value
+            tuple can contain up to 4 elements: (value, type, default, units)
+            'int' and 'float' force types, 'num' will decide based on the value
+            'fixed' will return the value modified
+        Returns:
+            str: The human-readable friendly/readable result
+    """
+    units:str = ''
+    default:str = ''
+
+    if isinstance(val, tuple): # Handle a tuple of 1-4 elements: (value, type, default, units)
+        if len(val) > 1: type = val[1]
+        if len(val) > 2: default = val[2]
+        if len(val) > 3: units = val[3]
+        if len(val) > 0: value = val[0]
+    else:
+        value:int|float|str|bool = val
+        if (isinstance(value, str) and re.match(value, r"^\d+-\d+-\d+ \d+\:\d+")): type = 'datetime'
+        if isinstance(value, bool): type = 'bool'
+        if isinstance(value, int) or isinstance(value, float): type = 'num'
+
+    # Fixed is left entirely alone
+    if type == 'fixed': return str(value) + units
+
+    # Empty, zero or false we return the default so the display isn't full of "No" and "0" etc.
+    if value in [None, False, 'False', 'false', 'NO', 'No', 'no', 0, '0', '', ' ', 'Null', 'null']: return default
+
+    ret:str = ""
+    match type:
+        case 'bool': # We're going to display Yes (blanks and False are handled above)
+            ret = "Yes"
+
+        case 'datetime': # If it's a datetime convert it from the json date format to our date format
+            ret = datetime.strptime(str(value), "%Y-%m-%d %H:%M:%S").strftime("%Y-%m-%d %H:%M")
+
+        case 'interval': # Approximated interval (no seconds, only show minutes if it's less than a day)
+            days , rem = divmod(int(value), 60*60*24)
+            hours, rem = divmod(rem, 60*60)
+            mins, rem = divmod(rem, 60)
+            tmp:list = []
+            if floor(days) > 1: tmp.append(f"{floor(days)} days")
+            elif int(days) > 0: tmp.append(f"1 day")
+            if floor(hours) > 1: tmp.append(f"{floor(hours)} hours")
+            elif int(hours) > 0: tmp.append(f" 1 hour")
+            if len(tmp) < 2:
+                if floor(mins) > 1: tmp.append(f" {int(mins)} minutes")
+                elif mins > 0: tmp.append(f" 1 minute")
+            ret = ' '.join(tmp)
+
+        case 'num' | 'float' | 'int': # We only shorten/simplify numbers over 100k. Smaller ones we just display with commas at thousands
+            if float(value) > 10000:
+                abbrs:list[str] = ['', 'K', 'M', 'B', 'T']  # Abbreviations for thousands, millions, billions, trillions
+                fnum:float = float('{:.3g}'.format(value))
+                magnitude = 0
+                while abs(fnum) >= 1000:
+                    if magnitude >= len(abbrs) - 1: break
+                    magnitude += 1
+                    fnum /= 1000.0
+                ret = '{}{}'.format('{:f}'.format(fnum).rstrip('0').rstrip('.'), abbrs[magnitude])
+            elif float(value) > 100 or type == 'int': # No decimals above 100
+                ret = f"{value:,.0f}"
+            elif float(value) > 10: # Only 1 above 10
+                ret = f"{value:,.1f}"
+            elif type == 'float': # Two if it's <10 and a float.
+                ret = f"{value:,.2f}"
+            else:
+                ret = f"{value:,}"
+
+        case _: # Title case two words, leave longer strings as is
+            ret = str(value).title() if str(value).count(' ') < 2 and re.search(r"[A-Z0-9]", str(value)) == None else str(value)
+
+    return ret + units
+
+def str_truncate(s:str, length:int = 20, elipsis:str = '…', loc:str = 'right') -> str:
+    """ Truncate a string to a specified length, adding an ellipsis if the string is longer than the specified length. """
+    if len(s) <= length:
+        return s
+
+    match loc:
+        case 'left':
+            return elipsis + s[-(length - len(elipsis)):]
+        case 'middle':
+            half_length = (length - len(elipsis)) // 2
+            return s[:half_length] + elipsis + s[-half_length:]
+        case _:
+            # Default to truncating at the right side
+            return s[:length - len(elipsis)] + elipsis
+
+class PopupNotice:
+    """ Create a temporary popup window """
+
+    def __init__(self, notice:str = '', timeout:int = 0, config = None) -> None:
+        self.config = config
+
+        self.root = tk.Tk()
+
+        if os.environ.get('XDG_SESSION_TYPE', 'x11').lower() == 'wayland':
+            print("Wayland detected: Using window types for borderless effect.")
+            # 'splash' or 'tooltip' usually removes decorations in Wayland
+            self.root.attributes('-type', 'splash')
+        else:
+            self.root.overrideredirect(True)
+
+        self.root.attributes("-alpha", 0.6)
+        self.root.geometry("350x150-1+0")
+        self.root.attributes("-topmost", True)
+
+        self.frame = tk.Frame(self.root, bg='red4', relief="raised")
+        self.frame.pack(fill="both", expand=True)
+
+        label = tk.Label(self.frame, text=notice, fg="white", bg="red4", font=("Helvetica", 12, "bold"), justify=tk.CENTER)
+        label.pack(pady=20, anchor=tk.CENTER)
+        exit_btn = tk.Button(self.frame, text="Close", fg="white", bg="red4", command=self.close)
+        exit_btn.pack(pady=10)
+
+        if timeout > 0: self.root.after(timeout, self.close)
+        self.frame.bind("<Button-1>", self.start_move)
+        self.frame.bind("<B1-Motion>", self.do_move)
+
+    def start_move(self, event) -> None:
+        self.x:int = event.x
+        self.y:int = event.y
+
+    def do_move(self, event) -> None:
+        deltax:int = event.x - self.x
+        deltay:int = event.y - self.y
+        x:int = self.root.winfo_x() + deltax
+        y:int = self.root.winfo_y() + deltay
+        self.root.geometry(f"+{x}+{y}")
+
+    @catch_exceptions
+    def close(self) -> None:
+        if self.root and self.root.winfo_exists():
+            #self.config.window_geometries['Alert'] = self.root.winfo_geometry()
+            self.root.destroy()